+++ wt_shortcodes/shortcodes/elements.php
@@ -9,7 +9,7 @@
  */
 
 /**
- * RECENT POSTS
+ * SEARCHFORM
  * ----------------------------------------------------------------
  * Print search form on the screen.
  */
@@ -17,18 +17,19 @@ if (!function_exists('wt_searchform')) {
 
 	function wt_searchform( $atts ) {
 
- 	    extract(shortcode_atts(array(
-	   		'default'		=> 'true',
-	   		'button-text' 	=> 'Search',
-	   	), $atts));
+		extract(shortcode_atts(array(
+			'button_text' 	=> 'Search',
+		), $atts));
 
- 	    // The custom search form
- 	    $output =  "<form role='search' id='wt_searchform'>";
- 	    $output .= "</form>";
-
+		// The custom search form
+		$output =  	"<div class='wt_searchform'>";
+		$output .= 		get_search_form( $echo );
+		$output .= 	"</div>";
 
 		return $output;
+
 	}
 	
 	add_shortcode('search', 'wt_searchform');
+	
 }